fix uint64 Nlink

+++ openwrt-feed/liquid-formula/src-subscription-gateway/gateway_disk_gc.go
@@ -463,7 +463,7 @@ func diskGCFileIdentityFromStat(
 		device: uint64(stat.Dev),
 		inode:  stat.Ino,
 		mode:   stat.Mode,
-		nlink:  stat.Nlink,
+		nlink:  uint64(stat.Nlink),
 		size:   stat.Size,
 		digest: digest,
 	}
